refactor(notifications): Replace SendNotification with a render primitive.
Request: Decide the framework-boundary question for notifications — the user
chose option B: the framework renders, apps and modules own the pipeline.
Motivation: The notificationAdapter existed only because the send pipeline was
framework JS — every orchestration op (insert, dedup, send, update) is an
ordinary YAML request, and a framework pipeline competes with Lowdefy's own
modules layer.

Decisions:
- RenderNotification renders one data item per call and returns
  { subject, title, preview, html, text, data } — arrays error with a pointer
  to the :for control; data is the link-resolved item so callers can render
  and store without re-deriving links.
- Link resolution stays in the step: serverUrl, landingPage and recordId are
  step properties (app.serverUrl / app.notificationLandingPage deleted, no
  VERCEL_URL fallback) — URL composition guards the rendered output and would
  be miserable as YAML string operators.
- landingPage without recordId errors when links exist: silent dead landing
  links are worse than a build-time correction.
- email.send capability deleted from SMTP/SendGrid: getMailSend was its only
  consumer; requests share the same send() and delivery filter.
- connection-mongodb changeset bump dropped: only test infra remains there.

Changes:
- api handleRenderNotification.js: renamed from handleSendNotification, now
  render-only (validate → resolve links → interpolate → schema → render).
- api derivePreview.js: preview derivation moved out of the deleted
  createNotificationRecord.js; getMailSend.js/getNotificationAdapter.js deleted.
- build lowdefySchema.js: app.serverUrl/notificationLandingPage removed;
  notification config loses emailConnectionId/dataConnectionId/delivery.
- build buildNotifications.js: connection-ref and delivery validation deleted.
- build validateRenderNotificationSteps.js: renamed validator, same logic.
- build validateStep.js/setStepId.js/countStepTypes.js: step renamed, literal
  array data now a build error; notification: id prefix kept.
- connection-mongodb MongoDBCollection.js: notificationAdapter deleted.
- connection-smtp SMTP.js, connection-sendgrid SendGridMail.js: email.send
  capability blocks removed.
- .changeset/feat-email-notifications.md: rewritten for the render primitive.

Tags: notifications, render-notification, framework-boundary, email,
notification-adapter, smtp, sendgrid, teardown, primitives

$30m
