+++ packages/api/src/routes/notifications/derivePreview.js
@@ -43,50 +43,4 @@ function derivePreview({ properties }) {
   return null;
 }
 
-// Record fields are snake_case: they are app-facing data — app pipelines
-// produce the conventional fields and app YAML queries the record.
-function createNotificationRecord({
-  id,
-  notificationConfig,
-  item,
-  properties,
-  contact,
-  email,
-  isValidEmail,
-  html,
-  text,
-  appName,
-}) {
-  return {
-    _id: id,
-    key: item.deduplication_key ?? null,
-    type: notificationConfig.notificationId,
-    template: notificationConfig.type,
-
-    contact_id: contact._id,
-    contact,
-    email,
-    is_valid_email: isValidEmail,
-
-    subject: properties.subject,
-    title: properties.title ?? properties.subject,
-    preview: derivePreview({ properties }),
-    body: html,
-    text,
-    data: item,
-
-    send_email: item.send_email ?? true,
-    cc: type.isArray(item.cc) ? item.cc : null,
-    bcc: type.isArray(item.bcc) ? item.bcc : null,
-    sent: false,
-    send_attempts: 0,
-    last_attempt: null,
-    email_result: null,
-
-    read: false,
-
-    created: { timestamp: new Date(), app_name: appName ?? null },
-  };
-}
-
-export default createNotificationRecord;
+export default derivePreview;
